Rename hallOfFameSlice
