+++ src/redux/store.js
@@ -1,6 +1,6 @@
 import { configureStore, combineReducers } from '@reduxjs/toolkit';
 import logger from 'redux-logger';
-import hallOfFameSlice from './hof/temp';
+import hallOfFameSlice from './hof/hallOfFameSlice';
 import justReleasedSlice from './just_released/justReleasedSlice';
 import popularSlice from './popular/popularSlice';
 import upcomingSlice from './upcoming/upcomingSlice';
